Individual plotting function created
Version of script saved with the ability to plot individual populations - not currently functional from the command line, as the plots are not output. 
Several blocks of code have been commented out - these are being retained on Github so they can be added back in, if the function-based approach cannot be troubleshot.