Update visualize_admixture.R
First functional version of the script.
The "pop_levels"/pop_order variable is hardcoded to the houbara paper, and will need to be altered. It should come as a text file argument and be supplied from the command line.

The patchwork plotting sections are also hardcoded, and do not produce the desired plots. We will need to alter the plot aesthetics and figure out how to supply individual plots directly to patchwork from a list object.