Add a mapping summary script
I wrote a quick-and-dirty bash script to summarize mapping stats into a TSV. It relies on Qualimap and Fastp, and is designed to integrate in with my other bash script (map.sh), which runs/attempts to run Qualimap and Fastp. I am storing it here so it doesn't get lost somewhere.