Create map.sh
This is our new OVH mapping script. It can stand alone, but is meant to integrate with 'get_map_stats.sh' and 'map_driver.sh'. It attempts to fastp-rawdata, maps with bwa and samtools, and then runs bamqc.

Known limitations are that if you attempt to fastp files which have already been fastp'd, it will erase them. It is also tied to the genome_assembly conda environment, and the conda environments in general are hardcoded.